Alert the user of invalid rights_statements (#520)
* Alert the user of invalid rights_statements

When a user uploads a CSV, check that all of the rights_statement values
are legit. Alert the user up front if they aren't.

* Do not titleize headers in warning messages

It is going to be confusing for users if their CSV header is called
'content_type' but the error message refers to 'Content Type'. Let's
keep things consistent for clarity.

* CI fix

* Metadata problems are warnings, not errors