Add the initial yellowback preprocessor framing
We've decided to keep the logic for collection specific processing
separate from the importer. This lets us minimize collection-sepecific
code in the CSV Importer.

The YellowbackPreprocessor will encapsulate all of the metadata and file
processing rules specific to the Yellowback collection and output a csv
manifest file suitable for ingest using the standard CSV Importer.