Add deduplication key attribute
This is a system-style attribute that is only meant
to be used by the importer to deduplicate rows. This
will make it easier to know which records need to be updated
or not when re-running imports. The existing dedup field
will not be present in all collections, but this one will.