Rename the Yellowback preprocessor spec file
The spec is getting kinda long and we'll need to add a number of new tests to
handle file mappings.  Rather than increasing the spec file size, we'll split the
tests between metadata and file related tests.  Since this test focuses on metadata,
we'll denote that in the spec name.