Provide a deduplication_key for yellowback manifests
Use the emory_ark as the deduplication key for yellowbacks since the other
fields likely to be unique like the MMSID can be duplicated when more than
one copy of the same yellowback volume exists. The ARK is the only ID that
is unique for each copy of yellowback.  See "The common moths of England" in
the test fixture for an example.