Add custom Emory FileSet related fields to guide
This adds several fields to the field guide that
are related to the custom file types that Emory uses
on `FileSet`s.

This also moves the `usage.yml` file from the `emory`
folder to the `zizia` folder. That's the default and
keeping it in the default location makes it easier to
test because you are able to load the correct vocab
without an override/initializer.

Connected to https://github.com/curationexperts/in-house/issues/407