Fix logic in failing AdminSetup test
The existing test counts the absolute number of admins created in the
test, but at some point the fixture introduced new admins.  I.E. the
total admin count in the test is the number of admins created by the
intial fixture PLUS the number created in the test.

This PR changes the test to check the delta during the test run instead
of the absolute value.