Generate IIIF image urls based on sha1
This changes the IIIF image builder lambda
so that it uses a seperate service to generate
the URL based on the sha1 stored in solr. The new
URL is used if the server has the `FEDORA_ADAPTER` env
variable set to 's3', otherwise it will generate the URLs
based on the `FileSet` id.

Connected to emory-libraries/dlp-lux#61