Adds fixity_check preservation_event
* Adds fixity_check preservation_event to fixity_check method
in FileSetFixityCheckService. Takes the `results` array which contains
results for fixity_checks on all files in a file_set and selects the
ones that failed. Then, we check if the new array for failures is
empty?; if empty, we create a success preservation_event; if not empty,
we create a failure preservation_event along with name of file and
the original sha1 value.