Fix a flapping admin setup test
The final AdminSetup spec adds three new admins, but also creates an
additional admin from the fixture file.  Depending on the order the tests
are randomized in, sometimes this fixture admin gets deleted before the spec
gets run and sometimes not, so the test intermittently fails when testing for
an absolute number of admins.  This commit refactors the test to check
that making three new admins increases the total admin count by 3 instead.