Hard code sidekiq workers (#881)
The env var SIDEKIQ_WORKERS is not being recognized by the ruby app. I've tried experimenting with dotenv and loading dotenv at different times in the application boot sequence and it still didn't work. The app has always been using the old default of 5 workers.

This PR rewrites the sidekiq.yml file so that we don't use SIDEKIQ_WORKERS

* Hard code sidekiq changes

* Increase timeouts for shutdown