Removes keyword arg from update_content
* This commit removes preferred keyword from update_content method call
in file_sets_controller. The preferred file symbol is being received as a
normal arg in file_set_actor's update_content method and we were previously
passing this a hashed key-value pair argument. By removing the key `preferred:`
from the file_sets_controller, we pass this as a normal argument instead of
the requirement of a keyword.