Manifests should always be available (#1009)
* Manifests should always be available

Move the IIIF manifest generation out of CurateGenericWorksController,
where it is gated by Shibboleth, and into the iiif controller, where it
is more publicly accessible. The goal is to let us load iiif images for
public_low_view objects.

* Refactor testing for manifest cacheing

Test the cache behavior against the ManifestBuilderService, which is
where the behavior now actually lives.

Also, parsing the JSON and testing for specific tags is less likely to
break than testing against a static fixture file.