Changes manifest_url to meet lux requirements
* We are changing the manifest url to `host_name/iiif/work.id/manifest`
to meet lux requirements.
* In this commit, we are also changing the way `build_manifest` gets
called. We now pass the presenter to the manifest builder class from
the works_controller so we don't need to worry about building a
presenter for the manifest builder service.
* We are also editing the iiif_controller's manifest method when it
calls the `build_manifest` method by passing a presenter that we
create in the controller itself. Here, when we create a presenter
for the work, we are passing the work's solr_doc, creating a new
ability for the `current_user`, and the request is nil. When the
request is nil, the manifest_url is being built by passing the
host manually to the polymorphic_url method.