Adds preferred_file sha1 to image url
* Adds sha1 for the preferred_file in the image url so that
UV can display the correct file. This new image url lives in the
manifest.
* Modifies the file_set_indexer to add an additional SOLR field
called `sha1_tesim` which saves sha1 values for all the
files in a file_set.
* Modifies the fetch_visibility method in the iiif_controller
to now fetch the visibility of the file_set using the `sha1_tesim`
field in SOLR instead of `digest_ssim`, this is because,
`digest_ssim` will always only have sha1 for the
preservation_master_file and the identifier passed as a param to
the iiif_controller will be the sha1 for the preferred_file.