Updates Bixby and resolves the conflicts
Capybara also needed to be upgraded. Bixby was upgraded to a commit
rather than a version to allow us to us Rails style private methods.

Uses `.find` rather than `.select ... .first`

Cleans up some ternary lines  This one is nice, because instead of
using a conditional, we can now simplify down to an OR.

Uses `.blank?` or `.present?`, and prefers `if` over `unless`