Redoes manifest generation
* We are redoing the manifest generation process for performance enhancement.
* The persistence of manifests is moved into its own background job. This way
we can show a placeholder manifest with text that asks users to wait for
manifest assembly.
* We are also changing the way manifests get cached. When indexing a work,
a MD5 hash is created for values that go into a manifest. A new manifest is
now generated for a work only if anyone or more of those values is edited.
When any value is edited a new MD5 hash is generated which we are now using
as the filename for manifests.